refactor!: create party bank account
Turns former supplier-specific feature into general util.
